Initial Stretch version - replaced source_code with  https://github.com/bruceg/vmailmgr

+++ lib/misc/maildir.cc
@@ -1,4 +1,4 @@
-// Copyright (C) 1999,2000 Bruce Guenter <bruceg@em.ca>
+// Copyright (C) 1999,2000 Bruce Guenter <bruce@untroubled.org>
 //
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -15,6 +15,7 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include <config.h>
+#include <errno.h>
 #include "maildir.h"
 #include <unistd.h>
 #include <string.h>
@@ -36,7 +37,9 @@ int mkdirp(const mystring& dirname, mode_t mode)
     if(mkdirp(dirname.left(i), 0755))
       return -1;
   }
-  return mkdir(dirname.c_str(), mode);
+  if(i != dirname.length() - 1)
+    return mkdir(dirname.c_str(), mode);
+  return 0;
 }
 
 bool make_maildir(const mystring& dirname)
@@ -44,18 +47,18 @@ bool make_maildir(const mystring& dirname)
   if(mkdirp(dirname, 0700))
     return false;
   mystring curdir = dirname + "/cur";
-  if(mkdir(curdir.c_str(), 0755)) {
+  if(mkdir(curdir.c_str(), 0755) && (errno != EEXIST)) {
     rmdir(dirname);
     return false;
   }
   mystring newdir = dirname + "/new";
-  if(mkdir(newdir.c_str(), 0755)) {
+  if(mkdir(newdir.c_str(), 0755) && (errno != EEXIST)) {
     rmdir(curdir);
     rmdir(dirname);
     return false;
   }
   mystring tmpdir = dirname + "/tmp";
-  if(mkdir(tmpdir.c_str(), 0755)) {
+  if(mkdir(tmpdir.c_str(), 0755) && (errno != EEXIST)) {
     rmdir(newdir);
     rmdir(curdir);
     rmdir(dirname);
@@ -66,28 +69,35 @@ bool make_maildir(const mystring& dirname)
 
 bool delete_directory(const mystring& dirname)
 {
-  DIR* dir = opendir(dirname.c_str());
-  if(!dir)
-    return false;
-  while(dirent* entry = readdir(dir)) {
-    const char* name = entry->d_name;
-    if(name[0] == '.' &&
-       (NAMLEN(entry) == 1 ||
-	(name[1] == '.' && NAMLEN(entry) == 2)))
-      continue;
-    mystring fullname = dirname + "/";
-    fullname += mystring(name, NAMLEN(entry));
-    if(is_dir(fullname.c_str())) {
-      if(!delete_directory(fullname)) {
+  int retry;
+  retry = 0;
+  do {
+    DIR* dir = opendir(dirname.c_str());
+    if(!dir)
+      return false;
+    while(dirent* entry = readdir(dir)) {
+      const char* name = entry->d_name;
+      if(name[0] == '.' &&
+	 (NAMLEN(entry) == 1 ||
+	  (name[1] == '.' && NAMLEN(entry) == 2)))
+	continue;
+      mystring fullname = dirname + "/";
+      fullname += mystring(name, NAMLEN(entry));
+      if(is_dir(fullname.c_str())) {
+	if(!delete_directory(fullname)) {
+	  closedir(dir);
+	  return false;
+	}
+      }
+      else if(unlink(fullname.c_str())) {
 	closedir(dir);
 	return false;
       }
     }
-    else if(unlink(fullname.c_str())) {
-      closedir(dir);
-      return false;
-    }
-  }
-  closedir(dir);
-  return !rmdir(dirname);
+    closedir(dir);
+    if (rmdir(dirname) == 0)
+      return true;
+    ++retry;
+  } while (retry < 3);
+  return false;
 }
